Making results scrollable
Previously results would not scroll and extend past the edge of
the page. This has been fixed to make them scroll within the limits
of the container.

Added check to see if the option is completely visible within the
viewport, if not it will scroll to selected item.

Fixes bug Make results scrollable #2